pyproject.toml
[project]
name = "contextual-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "contextual-client>=0.5.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
]
